--- .sonarcloud.properties
@@ -1,1 +1,0 @@
-
