+++ subprojects/jris-core/src/test/java/com/gmail/gcolaianni5/jris/usage/JRisUsageTest.java
@@ -8,8 +8,8 @@ import java.util.Collections;
 import java.util.List;
 
 import com.gmail.gcolaianni5.jris.JRis;
-import com.gmail.gcolaianni5.jris.RisRecord;
-import com.gmail.gcolaianni5.jris.RisType;
+import com.gmail.gcolaianni5.jris.domain.RisRecord;
+import com.gmail.gcolaianni5.jris.domain.RisType;
 import io.reactivex.Observable;
 import org.junit.jupiter.api.Test;
