+++ subprojects/jris-io/src/integration-test/kotlin/com/gmail/gcolaianni5/jris/JRisIOIntegrationTest.kt
@@ -1,5 +1,7 @@
 package com.gmail.gcolaianni5.jris
 
+import com.gmail.gcolaianni5.jris.domain.RisRecord
+import com.gmail.gcolaianni5.jris.domain.RisType
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.amshove.kluent.shouldContainAll
 import org.amshove.kluent.shouldHaveSize
