+++ subprojects/jris-io/src/integration-test/kotlin/com/gmail/gcolaianni5/jris/usage/JRisIoUsageImportSpec.kt
@@ -1,7 +1,7 @@
 package com.gmail.gcolaianni5.jris.usage
 
-import com.gmail.gcolaianni5.jris.RisRecord
-import com.gmail.gcolaianni5.jris.RisType
+import com.gmail.gcolaianni5.jris.domain.RisRecord
+import com.gmail.gcolaianni5.jris.domain.RisType
 import com.gmail.gcolaianni5.jris.accept
 import com.gmail.gcolaianni5.jris.process
 import kotlinx.coroutines.ExperimentalCoroutinesApi
