+++ subprojects/jris-io/src/main/kotlin/com/gmail/gcolaianni5/jris/JRisIO.kt
@@ -1,5 +1,6 @@
 package com.gmail.gcolaianni5.jris
 
+import com.gmail.gcolaianni5.jris.domain.RisRecord
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.asFlow
@@ -32,7 +33,8 @@ object JRisIO {
     @ExperimentalCoroutinesApi
     @Throws(IOException::class)
     fun process(reader: Reader): List<RisRecord> = runBlocking(Dispatchers.IO) {
-        BufferedReader(reader).readLines().asFlow().toRisRecords().toList()
+        val lineFlow = BufferedReader(reader).readLines().asFlow()
+        JRis.process(lineFlow).toList()
     }
 
     /**
@@ -131,75 +133,3 @@ object JRisIO {
 
     //endregion
 }
-
-//region Kotlin extension functions for importing JRis files
-
-/**
- * Converts the RISFile lines provided by the reader as receiver into a list of [RisRecord]s.
- * May throw an [IOException] if the reader fails to deliver lines or a [JRisException]
- * if the lines cannot be parsed successfully.
- */
-@ExperimentalCoroutinesApi
-fun Reader.process() = JRisIO.process(this)
-
-/**
- * Converts the RISFile lines in the [File] provided as receiver into a list of [RisRecord]s.
- * May throw an [IOException] if the file cannot be read successfully.
- * or a [JRisException] if the lines cannot be parsed successfully.
- */
-@ExperimentalCoroutinesApi
-fun File.process() = JRisIO.process(this)
-
-/**
- * Converts the RISFile lines from the file with the path provided as receiver into a list of [RisRecord]s.
- * May throw an [IOException] if the file cannot be read successfully.
- * or a [JRisException] if the lines cannot be parsed successfully.
- */
-@ExperimentalCoroutinesApi
-fun String.process() = JRisIO.process(this)
-
-/**
- * Converts the RISFile lines provided by the [InputStream] as receiver  into a list of [RisRecord]s.
- * May throw an [IOException] if the stream cannot be read successfully.
- * or a [JRisException] if the lines cannot be parsed successfully.
- */
-@ExperimentalCoroutinesApi
-fun InputStream.process() = JRisIO.process(this)
-
-//endregion
-
-//region Kotlin extension functions for exporting JRis files
-
-/**
- * Converts a list of [RisRecord]s into a list of [String]s in RIS file format, dumping them into the
- * [Writer] provided as receiver. Optionally accepts a list of names of [RisTag]s defining a sort order for
- * the [RisTag]s in the file.
- */
-@ExperimentalCoroutinesApi
-fun Writer.accept(records: List<RisRecord>, sort: List<String> = emptyList()) = JRisIO.export(records, sort, this)
-
-/**
- * Converts a list of [RisRecord]s into a list of [String]s in RIS file format, writing them into the [File]
- * provided as receiver. Optionally accepts a list of names of [RisTag]s defining a sort order for the [RisTag]s
- * in the file.
- */
-@ExperimentalCoroutinesApi
-fun File.accept(records: List<RisRecord>, sort: List<String> = emptyList()) = JRisIO.export(records, sort, this)
-
-/**
- * Converts a list of [RisRecord]s into a list of [String]s in RIS file format, writing them into
- * the [OutputStream] provided as receiver. Optionally accepts a list of names of [RisTag]s defining a sort order
- * for the [RisTag]s in the file.
- */
-@ExperimentalCoroutinesApi
-fun OutputStream.accept(records: List<RisRecord>, sort: List<String> = emptyList()) = JRisIO.export(records, sort, this)
-
-/**
- * Converts a list of [RisRecord]s into a list of [String]s in RIS file format, writing them into file with
- * the path provided as the receiver, if possible.
- * Optionally accepts a list of names of [RisTag]s defining a sort order for the [RisTag]s in the file.
- */
-@ExperimentalCoroutinesApi
-fun String.accept(records: List<RisRecord>, sort: List<String> = emptyList()) = JRisIO.export(records, sort, this)
-
-//endregion
